First cut of import almost complete (error reprocessing remains)

+++ BioLink.Client.Extensibility/FormattedLatLogVirtualColumn.cs
@@ -91,7 +91,7 @@ namespace BioLink.Client.Extensibility {
 
         }
 
-        private string FormatDate(long d, bool asRomanMonth = false) {
+        private string FormatDate(int d, bool asRomanMonth = false) {
             if (d == 0) {
                 return "";
             }
